Added -openResultSetQueries method, for debugging
If you get a runtime error from sqlite saying it can't do something
(close the database, or vacuum) because there are still queries in
progress, it probably means you forgot to call -close on a FMResultSet.
Calling this method will return the SQL queries of all open results
sets, making it a lot easier to figure out where in your source code
the problem is.
(I didn't add this to the header because it's not intended to be
called programmatically, rather by the developer from gdb.)